pyproject.toml
[tool.poetry]
name = "command-line-functions"
version = "0.1.0"
description = ""
authors = ["Kean Mattingly <kean.mattingly@equipmentshare.com>"]
readme = "README.md"
packages = [
	{include = "command_line_functions"},
	{include = "db_auth"},
	{include = "open_git_in_browser"},
]

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.9.0"}
gitpython = "^3.1.40"

[tool.poetry.scripts]
dbAuth = "db_auth.main:app"
og = "open_git_in_browser.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
